requirements.txt
pandas
streamlit
openai>=1.0.0
python-dotenv
